Fix build failures on solaris and PPC for bit shift alignment and size_t -> int implicit conversion

+++ plugin/command_log/command_log.cc
@@ -187,7 +187,7 @@ void CommandLog::apply(const message::Command &to_apply)
 
   /* We always write in network byte order */
   unsigned char nbo_length[8];
-  int8store(nbo_length, length);
+  int8store(nbo_length, (uint64_t) length);
 
   /* Write the length header */
   do
